refactor: split api/standards_routes.py into focused modules
